requirements.txt
# Streamlit App Requirements
streamlit>=1.29.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=10.0.0

# Optional: If your notebooks use these
# scikit-learn>=1.3.0
# xgboost>=2.0.0
# seaborn>=0.12.0
